+++ src/index.tsx
@@ -6,18 +6,18 @@ import App from "./App";
 import Home from "./components/Layout/Home";
 import Register from "./components/Login/Register";
 import Login from "./components/Login/Login";
-import Contact from "./components/Contact";
-import Footer from "./components/Footer";
-import OurTeam from "./components/OurTeam";
+import Contact from "./components/Layout/Contact";
+import Footer from "./components/Layout/Footer";
+import OurTeam from "./components/Layout/OurTeam";
 import SearchAI from "./components/RecipeAI/SearchAI";
 import SearchChoice from "./components/Search/SearchChoice";
 import SavedRecipes from "./components/SavedRecipes/SavedRecipes";
 import Layout from "./components/Layout/Layout";
 import SingleRecipePage from "./components/SingleRecipePage/SingleRecipePage";
-import EditRecipe from "./components/EditRecipe";
-import Planner from "./components/Planner";
-import ShoppingList from "./components/ShoppingList";
-import NotFound from "./components/NotFound";
+import EditRecipe from "./components/EditRecipe/EditRecipe";
+import Planner from "./components/Planner/Planner";
+import ShoppingList from "./components/ShoppingList/ShoppingList";
+import NotFound from "./components/Layout/NotFound";
 import theme from "./CustomTheme";
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
